Product production and bug fix

+++ src/main/java/com/example/ConfectionerWebsite/services/ProductionService.java
@@ -23,7 +23,7 @@ public class ProductionService {
     public void callProduceProductProcedure(Long finishedProductId, Double quantity, Long employeeId) {
         entityManager.createNativeQuery("CALL produce_product(:fp_id, :qty, :emp_id)")
                 .setParameter("fp_id", finishedProductId)
-                .setParameter("qty", quantity.intValue()) // Преобразуем к int, если нужно
+                .setParameter("qty", quantity.doubleValue()) // Преобразуем к int, если нужно
                 .setParameter("emp_id", employeeId)
                 .executeUpdate();
     }
@@ -31,5 +31,4 @@ public class ProductionService {
     public List<ProductProduction> getAllProductions(){
         return productionRepository.findAll();
     }
-
 }
